WIP: restructure file tree
