+++ comiccrawler/download_manager.py
@@ -14,14 +14,18 @@ from time import time
 
 from worker import Worker, current, await_
 
+from .analyzer import analyze
 from .safeprint import print
+from .batch_analyzer import BatchAnalyzer
 from .config import setting
-from .core import download, analyze, safefilepath, BatchAnalyzer, create_mission
+from .core import download, create_mission
 from .profile import get as profile
 from .error import PauseDownloadError
+from .util import safefilepath
 
-from .mission_manager import mission_manager, load_episodes
+from .mission_manager import mission_manager
 from .channel import download_ch
+from .episode_loader import load_episodes
 
 def quote(item):
 	if sys.platform == "win32":
